added invoice service

+++ gamestore/src/main/java/com/company/gamestore/model/Fee.java
@@ -1,6 +1,7 @@
 package com.company.gamestore.model;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Objects;
 
 @Entity
@@ -9,7 +10,7 @@ public class Fee {
 
     @Id
     private String productType;
-    private Double fee;
+    private BigDecimal fee;
 
     public String getProductType() {
         return productType;
@@ -19,11 +20,11 @@ public class Fee {
         this.productType = productType;
     }
 
-    public Double getFee() {
+    public BigDecimal getFee() {
         return fee;
     }
 
-    public void setFee(Double fee) {
+    public void setFee(BigDecimal fee) {
         this.fee = fee;
     }
